requirements.txt
click==8.0.3
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
PyMySQL==1.0.2
SQLAlchemy==1.4.29
Werkzeug==2.2.2
requests==2.31.0
Pillow==10.3.0
volcengine-python-sdk==4.0.2